requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
requests==2.28.1
html5lib
lxml
flask-session
psutil
flask-sse
eventlet
mysql-connector-python
python-dotenv
gunicorn
Authlib
flask-mail
razorpay
flask_login